+++ packages/extension-polkagate/src/popup/passwordManagement/constants.ts
@@ -1,0 +1,10 @@
+// Copyright 2019-2023 @polkadot/extension-polkagate authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+export const STEPS = {
+  NO_PASSWORD: 0,
+  ALREADY_SET_PASSWORD: 1,
+  NEW_PASSWORD_SET: 2,
+  PASSWORD_REMOVED: 3,
+  ERROR: 5
+}
